+++ package.json
@@ -49,6 +49,7 @@
   "devDependencies": {
     "@babel/core": "^7.11.4",
     "@expo/babel-preset-cli": "^0.2.17",
+    "@expo/config-plugins": "~1.0.29",
     "@expo/xdl": "^57.9.31",
     "@types/jest": "^26.0.10",
     "@types/json-schema-traverse": "^0.4.0",
@@ -69,10 +70,7 @@
     "vsce": "^1.79.3",
     "vscode-test": "^1.4.0",
     "webpack": "^4.44.1",
-    "webpack-cli": "^3.3.12"
-  },
-  "dependencies": {
-    "@expo/config": "^3.3.38",
+    "webpack-cli": "^3.3.12",
     "find-up": "^5.0.0",
     "jsonc-parser": "^3.0.0"
   }
