+++ src/manifest/utils/async.ts
